fix bug

+++ app/Http/Controllers/OrderController.php
@@ -565,7 +565,10 @@ class OrderController extends Controller
             $order_detail = OrderDetail::where('products_id', $products_id)->where('orders_id', $order->id)->first();
             if ($order_detail) {
                 OrderVideoDetail::where('order_details_id', $order_detail->id)->where('product_details_videos_id', $product_detail_videos_id)->delete();
-                $order_detail->delete();
+                $order_video_detail = OrderVideoDetail::where('order_details_id', $order_detail->id)->first();
+                if($order_video_detail == null) {
+                    $order_detail->delete();
+                }
                 UserProduct::where('users_id', $users_id)->where('products_id', $products_id)->where('partial', 1)->delete();
                 UserVideoSession::where('users_id', $users_id)->where('video_sessions_id', $product_detail_video->video_sessions_id)->delete();
                 $found_refund = Refund::where('users_id', $users_id)->where('products_id', $products_id)->where('orders_id', $order->id)->first();
